iranwellness-master/iranwellness-master/requirements.txt
bs4==0.0.1
Django==2.2.3
django-debug-toolbar==3.2
django-jalali==4.1.2
django-filter==2.4.0
djangorestframework==3.12.2
jdatetime==3.6.2
kavenegar==1.1.2
persiantools==1.4.1
Pillow==6.2.1
psycopg2-binary==2.8.3
pytz==2019.1
requests==2.22.0
sentry-sdk==0.19.5
urllib3==1.25.3
zeep==3.4.0